pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "dynamite"
version = "0.1.0"
description = "DynaMITE: Dynamic Mismatch Inference via Transformer Encoder for Robust Humanoid Locomotion"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "DynaMITE Authors"},
]

dependencies = [
    "torch>=2.1.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "tensorboard>=2.14.0",
    "matplotlib>=3.7.0",
    "scipy>=1.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
]
analysis = [
    "scikit-learn>=1.3.0",
    "pandas>=2.0.0",
    "seaborn>=0.12.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"

[tool.ruff]
line-length = 100
target-version = "py310"
